requirements.txt
requests
openai
gspread
google-auth